% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{actions_block}
\alias{actions_block}
\title{Actions Block}
\usage{
actions_block(elements, id = NULL)
}
\arguments{
\item{elements}{An array of interactive element objects:
buttons, select menus, overflow menus, or date pickers.
There is a maximum of 5 elements in each action block.}

\item{id}{character, A string acting as a unique identifier for a block.
Default: NULL}
}
\value{
actions_block class
}
\description{
A block that is used to hold interactive elements.
}
\details{
Available in surfaces: \emph{Modals}, \emph{Messages}, \emph{Home tabs}
}
\examples{

# An actions block with a select menu and a button:

actions_block(
  id = 'actions1',
  elements = wrap_blocks(
    select_menu(
      placeholder = 'Which witch is the witchiest witch?',
      id = 'select_2',
      options = as.blocks(
        lapply(c('Matilda','Glinda','Granny Weatherwax','Hermione'),
        FUN = function(x){
          block_option(
            text = x, 
            value = gsub('\\\\s','',tolower(x))
           )
        })
     )
    ),
    button_element(
      text = 'Cancel',
      value = 'cancel',
      id = 'button_1'
    )
 )
)

# An actions block with a datepicker, an overflow, and a button:

actions_block(
  id = 'actionblock789',
  elements = wrap_blocks(
    datepicker_element(
      id = 'datepicker123',
      initial_date = '1990-04-28',
      placeholder = 'Select a date'
    ),
    overflow_menu(
      id = 'overflow',
      options = as.blocks(
        lapply(sprintf('value-\%d',1:4),block_option,
        text = '*this is plain_text text*')
       )
    ),
    button_element(
      text = 'Click Me',
      value = 'click_me_123',
      id = 'button'
    )
  )
)

}
\seealso{
Other layout: 
\code{\link{as.blocks}()},
\code{\link{context_block}()},
\code{\link{divider_block}()},
\code{\link{file_block}()},
\code{\link{image_block}()},
\code{\link{input_block}()},
\code{\link{section_block}()},
\code{\link{wrap_blocks}()}
}
\concept{layout}
