% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.block_text}
\alias{as.block_text}
\title{FUNCTION_TITLE}
\usage{
as.block_text(x, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
