% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{as.blocks}
\alias{as.blocks}
\title{FUNCTION_TITLE}
\usage{
as.blocks(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
