% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositions.R
\name{block_confirm}
\alias{block_confirm}
\title{FUNCTION_TITLE}
\usage{
block_confirm(title, text, confirm, deny)
}
\arguments{
\item{title}{PARAM_DESCRIPTION}

\item{text}{PARAM_DESCRIPTION}

\item{confirm}{PARAM_DESCRIPTION}

\item{deny}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
