% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{block_datepicker}
\alias{block_datepicker}
\title{FUNCTION_TITLE}
\usage{
block_datepicker(id, placeholder = NULL, initial_date = NULL,
  confirm = NULL)
}
\arguments{
\item{id}{PARAM_DESCRIPTION}

\item{placeholder}{PARAM_DESCRIPTION, Default: NULL}

\item{initial_date}{PARAM_DESCRIPTION, Default: NULL}

\item{confirm}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
block_datepicker(id = 'datepicker123',initial_date = "1990-04-28", 
placeholder = 'Select a date')
}
