% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{file_block}
\alias{file_block}
\title{File Block}
\usage{
file_block(external_id, id = NULL)
}
\arguments{
\item{external_id}{character, the external unique ID for this file.}

\item{id}{character, A string acting as a unique identifier for a block.
Default: NULL}
}
\value{
block_file class
}
\description{
Displays a \href{https://api.slack.com/messaging/files/remote}{remote file}
}
\details{
Available in surfaces: \emph{Messages}
}
\examples{
file_block(
  external_id = 'ABCD1'
)
}
\seealso{
Other layout: 
\code{\link{actions_block}()},
\code{\link{as.blocks}()},
\code{\link{context_block}()},
\code{\link{divider_block}()},
\code{\link{image_block}()},
\code{\link{input_block}()},
\code{\link{section_block}()},
\code{\link{wrap_blocks}()}
}
\concept{layout}
