% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositions.R
\name{block_option}
\alias{block_option}
\title{FUNCTION_TITLE}
\usage{
block_option(text, value, description = NULL, url = NULL)
}
\arguments{
\item{text}{PARAM_DESCRIPTION}

\item{value}{PARAM_DESCRIPTION}

\item{description}{PARAM_DESCRIPTION, Default: NULL}

\item{url}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
