% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositions.R
\name{block_option_groups}
\alias{block_option_groups}
\title{FUNCTION_TITLE}
\usage{
block_option_groups(opts)
}
\arguments{
\item{opts}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
