% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{block_radio_buttons}
\alias{block_radio_buttons}
\title{Radio button group element}
\usage{
block_radio_buttons(id, opts, initial_options = NULL, confirm = NULL)
}
\arguments{
\item{id}{character, An identifier for the action triggered
when the radio button group is changed.}

\item{opts}{list, a list of block_options.}

\item{initial_options}{An option_block that exactly
matches one of the options within options, Default: NULL}

\item{confirm}{block_confirm, defines an optional
confirmation dialog after the button is clicked. Default: NULL}
}
\value{
block_radio_buttons class
}
\description{
A radio button group that allows a
user to choose one item from a list of possible options.
}
\details{
\itemize{
\item Works with block types: \emph{Section}, \emph{Actions}, \emph{Input}
}
}
