% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositions.R
\name{block_text}
\alias{block_text}
\title{Create a block for text}
\usage{
block_text(
  text,
  type = c("mrkdwn", "plain_text"),
  emoji = NULL,
  verbatim = NULL
)
}
\arguments{
\item{text}{character, the text for the block.}

\item{type}{character, type of block, Default: c("mrkdwn", "plain_text")}

\item{emoji}{logical, Indicates whether emojis in a text field should be
escaped into the colon emoji format. This field is only usable
when type is plain_text. Default: NULL}

\item{verbatim}{When set to FALSE (Default value when NULL) URLs
will be auto-converted into links, conversation names will be
link-ified, and certain mentions will be automatically parsed.
Default: NULL}
}
\value{
block_text class
}
\description{
Create slack block that contains text
}
\details{
For more details see \href{https://api.slack.com/reference/block-kit/composition-objects#text}{slack documentation}
}
\examples{
block_text('A message *with some bold text* and _some italicized text_.')
}
