% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{button_element}
\alias{button_element}
\title{Button element}
\usage{
button_element(
  id,
  text,
  value = NULL,
  url = NULL,
  style = NULL,
  confirm = NULL
)
}
\arguments{
\item{id}{character, An identifier for this action.}

\item{text}{character, text of the button}

\item{value}{The value to send along with the interaction
payload. Default: NULL}

\item{url}{character, A URL to load in the user's browser
when the button is clicked. Default: NULL}

\item{style}{Decorates buttons with alternative visual
color schemes, Default: NULL}

\item{confirm}{block_confirm, defines an optional
confirmation dialog after the button is clicked. Default: NULL}
}
\value{
block element class
}
\description{
An interactive component that inserts a button.
The button can be a trigger for anything from opening a
simple link to starting a complex workflow.
}
\details{
\itemize{
\item Works with block types: \emph{Section}, \emph{Actions}
\item \strong{primary} gives buttons a green outline and text, ideal for affirmation or confirmation actions.
\item \strong{primary} should only be used for one button within a set.
\item \strong{danger} gives buttons a red outline and text, and should be used when the action is destructive.
\item Use \strong{danger} even more sparingly than primary.
\item If you don't include this field, the \strong{default} button style will be used.
}
}
\examples{
# A regular interactive button:
button_element(text = 'Click Me', value = 'click_me_123', id = 'button')

# A button with a primary style attribute:
button_element(text = 'Click Me', value = 'click_me_123', id = 'button',style = 'primary')

# A link button:

button_element(text = 'Link Button', url = 'https://api.slack.com/block-kit', id = 'button')

}
\seealso{
Other elements: 
\code{\link{checkbox_element}()},
\code{\link{datepicker_element}()},
\code{\link{image_element}()},
\code{\link{overflow_menu}()},
\code{\link{plaintext_element}()},
\code{\link{radiobuttons_element}()},
\code{\link{select_menu}()},
\code{\link{select_type_menu}()}
}
\concept{elements}
