% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{checkbox_element}
\alias{checkbox_element}
\title{Checkbox groups}
\usage{
checkbox_element(id, options, initial_options = NULL, confirm = NULL)
}
\arguments{
\item{id}{character, an identifier for the action triggered when the
checkbox group is changed.}

\item{options}{An list of block_option objects.}

\item{initial_options}{An list of block_option objects that exactly
matches one or more of the options within options. Default: NULL}

\item{confirm}{A confirm object that defines an optional confirmation
dialog that appears after clicking one of the checkboxes in this element.
Default: NULL}
}
\value{
block element class
}
\description{
A checkbox group that allows a user to choose multiple
items from a list of possible options.
}
\details{
Works with block types: \emph{Section}, \emph{Actions}, \emph{Input}
}
\examples{
section_block(
text = 'Check out these charming checkboxes',
accessory = checkbox_element(
 id = 'my_checkbox',
 options = wrap_blocks(
 block_option(
   value = 'A1',
   text = 'Checkbox 1'
 ),
 block_option(
   value = 'A2',
   text = 'Checkbox 2'
 )
 ),
 initial_options = wrap_blocks(
   block_option(
     value = 'A1',
     text = 'Checkbox 1'
   )
 )
)
)
}
\seealso{
\code{\link[checkmate]{checkClass}}

Other elements: 
\code{\link{button_element}()},
\code{\link{datepicker_element}()},
\code{\link{image_element}()},
\code{\link{overflow_menu}()},
\code{\link{plaintext_element}()},
\code{\link{radiobuttons_element}()},
\code{\link{select_menu}()},
\code{\link{select_type_menu}()}
}
\concept{elements}
