% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{context_block}
\alias{context_block}
\title{Context Block}
\usage{
context_block(elements, id = NULL)
}
\arguments{
\item{elements}{An array of image elements and text objects. Maximum number of items is 10.}

\item{id}{character, A string acting as a unique identifier for a block.
Default: NULL}
}
\value{
context_block class
}
\description{
Displays message context, which can include both images and text.
}
\details{
Available in surfaces: \emph{Modals}, \emph{Messages}, \emph{Home tabs}
}
\examples{
context_block(
  elements = wrap_blocks(
    image_element(
      url = 'https://image.freepik.com/free-photo/red-drawing-pin_1156-445.jpg',
      alt_text = 'images'
    ),
    block_text(
      text = 'Location: **Dogpatch**'
    )
  )
)
}
\seealso{
Other layout: 
\code{\link{actions_block}()},
\code{\link{as.blocks}()},
\code{\link{divider_block}()},
\code{\link{file_block}()},
\code{\link{image_block}()},
\code{\link{input_block}()},
\code{\link{section_block}()},
\code{\link{wrap_blocks}()}
}
\concept{layout}
