% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{divider_block}
\alias{divider_block}
\title{Divider Block}
\usage{
divider_block(id = NULL)
}
\arguments{
\item{id}{character, unique identifier for a block. Default: NULL}
}
\value{
divider_block class
}
\description{
A content divider, like an hr HTML DOM,
to split up different blocks inside of a message.
}
\details{
Available in surfaces: \emph{Modals}, \emph{Messages}, \emph{Home tabs}
}
\examples{
divider_block()
}
\seealso{
Other layout: 
\code{\link{actions_block}()},
\code{\link{as.blocks}()},
\code{\link{context_block}()},
\code{\link{file_block}()},
\code{\link{image_block}()},
\code{\link{input_block}()},
\code{\link{section_block}()},
\code{\link{wrap_blocks}()}
}
\concept{layout}
