% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{image_block}
\alias{image_block}
\title{Image Block}
\usage{
image_block(image_url, alt_text, title = NULL, id = NULL)
}
\arguments{
\item{image_url}{character, The URL of the image to be displayed.}

\item{alt_text}{character, A plain-text summary of the image.
This should not contain any markup.}

\item{title}{character, an optional title for the image. Default: NULL}

\item{id}{character, unique identifier for a block. Default: NULL}
}
\value{
block_divider class
}
\description{
A simple image block, designed to make those cat photos really pop.
}
\details{
Available in surfaces: \emph{Modals}, \emph{Messages}, \emph{Home tabs}
}
\examples{
image_block(
 image_url = 'http://placekitten.com/500/500',
 alt_text = 'An incredibly cute kitten.',
 title = 'Please enjoy this photo of a kitten',
 id = 'image4'
)
}
\seealso{
Other layout: 
\code{\link{actions_block}()},
\code{\link{as.blocks}()},
\code{\link{context_block}()},
\code{\link{divider_block}()},
\code{\link{file_block}()},
\code{\link{input_block}()},
\code{\link{section_block}()},
\code{\link{wrap_blocks}()}
}
\concept{layout}
