% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{image_element}
\alias{image_element}
\title{Image element}
\usage{
image_element(url = NULL, alt_text = "image")
}
\arguments{
\item{url}{character, URL of the image to be displayed, Default: NULL}

\item{alt_text}{character, text summary of the image, Default: 'image'}
}
\value{
block element class
}
\description{
A simple image block
}
\details{
\itemize{
\item Works with block types: \emph{Section}, \emph{Context}
}
}
\examples{

image_element(
  url = 'http://placekitten.com/700/500',
  alt_text = 'Multiple cute kittens.'
)

}
\seealso{
Other elements: 
\code{\link{button_element}()},
\code{\link{checkbox_element}()},
\code{\link{datepicker_element}()},
\code{\link{overflow_menu}()},
\code{\link{plaintext_element}()},
\code{\link{radiobuttons_element}()},
\code{\link{select_menu}()},
\code{\link{select_type_menu}()}
}
\concept{elements}
