% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{input_block}
\alias{input_block}
\title{Input Block}
\usage{
input_block(label, element, id = NULL, hint = NULL, optional = NULL)
}
\arguments{
\item{label}{character, a label that appears above an input element.}

\item{element}{An plain-text input element, a select menu element,
a multi-select menu element, or a datepicker.}

\item{id}{character, A string acting as a unique identifier for a block.
Default: NULL}

\item{hint}{An optional hint that appears below an input element in a lighter grey.
Default: NULL}

\item{optional}{A boolean that indicates whether the input element may be
empty when a user submits the modal. Default: NULL}
}
\value{
input_block class
}
\description{
A block that collects information from users -
it can hold a plain-text input element, a select menu element,
a multi-select menu element, or a datepicker.
}
\details{
Available in surfaces: \emph{Modals}, \emph{Messages}, \emph{Home tabs}
}
\examples{
input_block(
  id = 'input123',
  label = 'Label of input',
  element = plaintext_element(
    id = 'plain_input',
    placeholder = 'Enter some plain text'
  )
)
}
\seealso{
Other layout: 
\code{\link{actions_block}()},
\code{\link{as.blocks}()},
\code{\link{context_block}()},
\code{\link{divider_block}()},
\code{\link{file_block}()},
\code{\link{image_block}()},
\code{\link{section_block}()},
\code{\link{wrap_blocks}()}
}
\concept{layout}
