% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.block_option}
\alias{is.block_option}
\title{FUNCTION_TITLE}
\usage{
is.block_option(x, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
