% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.block_text}
\alias{is.block_text}
\alias{as.block_text.block_text}
\alias{as.block_text.NULL}
\alias{as.block_text.character}
\title{FUNCTION_TITLE}
\usage{
is.block_text(x, ...)

\method{as.block_text}{block_text}(x, ...)

\method{as.block_text}{`NULL`}(x, ...)

\method{as.block_text}{character}(x, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
