% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{overflow_menu}
\alias{overflow_menu}
\title{Overflow menu element}
\usage{
overflow_menu(id, options, confirm = NULL)
}
\arguments{
\item{id}{character, An identifier for the action triggered when a menu
option is selected.}

\item{options}{An list of block_option objects.}

\item{confirm}{A confirm object that defines an optional confirmation
dialog that appears after a menu item is selected.}
}
\value{
block menu class
}
\description{
This is like a cross between a button and a select menu when a
user clicks on this overflow button, they will be presented with a list of
options to choose from. Unlike the select menu, there is no typeahead field,
and the button always appears with an ellipsis ("…") rather than
customisable text.
}
\details{
Works with block types: \emph{Section}, \emph{Actions}
}
\examples{
section_block(
 id = 'section 890',
 text = block_text('This is a section block with an overflow menu.'),
 accessory =  overflow_menu(
      id = 'overflow',
      options = as.blocks(
        lapply(sprintf('value-\%d',1:4),block_option,
        text = '*this is plain_text text*')
       )
    )
)
}
\seealso{
Other elements: 
\code{\link{button_element}()},
\code{\link{checkbox_element}()},
\code{\link{datepicker_element}()},
\code{\link{image_element}()},
\code{\link{plaintext_element}()},
\code{\link{radiobuttons_element}()},
\code{\link{select_menu}()},
\code{\link{select_type_menu}()}
}
\concept{elements}
