% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{plaintext_element}
\alias{plaintext_element}
\title{Plain-text input element}
\usage{
plaintext_element(
  id,
  placeholder,
  initial_value = NULL,
  multiline = NULL,
  min_length = NULL,
  max_length = NULL
)
}
\arguments{
\item{id}{character, an identifier for the input value when the parent modal
is submitted.}

\item{placeholder}{character, A text object that defines the placeholder text
shown in the plain-text input.}

\item{initial_value}{character, the initial value in the plain-text
input when it is loaded. Default: NULL}

\item{multiline}{logical, indicates whether the input will be a single
line (FALSE) or a larger textarea (TRUE). Default: NULL}

\item{min_length}{integer, the minimum length of input that the user must
provide. If the user provides less, they will receive an error. Default: NULL}

\item{max_length}{integer, The maximum length of input that the user can provide.
If the user provides more, they will receive an error in the response. Default: NULL}
}
\value{
block element object
}
\description{
A plain-text input, similar to the HTML input tag,
creates a field where a user can enter freeform data.
It can appear as a single-line field or a larger textarea
using the multiline flag.
}
\details{
Works with block types: \emph{Section}, \emph{Actions}, \emph{Input}
}
\examples{
plaintext_element(
  id = 'plain_input',
  placeholder = 'Enter some plain text'
)
}
\seealso{
Other elements: 
\code{\link{button_element}()},
\code{\link{checkbox_element}()},
\code{\link{datepicker_element}()},
\code{\link{image_element}()},
\code{\link{overflow_menu}()},
\code{\link{radiobuttons_element}()},
\code{\link{select_menu}()},
\code{\link{select_type_menu}()}
}
\concept{elements}
