% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post_block}
\alias{post_block}
\alias{post_thread}
\title{FUNCTION_TITLE}
\usage{
post_block(block, channel, thread_ts = NULL,
  token = Sys.getenv("SLACK_API_TOKEN"))

post_thread(res, block)
}
\arguments{
\item{block}{PARAM_DESCRIPTION}

\item{channel}{character, Channel ID, label or a URL link to a message}

\item{thread_ts}{character slack api timestamp of the format
xxxxxxxxxx.xxxxxx, Default: NULL}

\item{token}{PARAM_DESCRIPTION, Default: Sys.getenv("SLACK_API_TOKEN")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[slackcalls]{post_slack}}
}
