% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post_thread}
\alias{post_thread}
\title{FUNCTION_TITLE}
\usage{
post_thread(res, block)
}
\arguments{
\item{res}{PARAM_DESCRIPTION}

\item{block}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
