% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{radio_buttons}
\alias{radio_buttons}
\title{FUNCTION_TITLE}
\usage{
radio_buttons(id, opts, initial_options = NULL, confirm = NULL)
}
\arguments{
\item{id}{PARAM_DESCRIPTION}

\item{opts}{PARAM_DESCRIPTION}

\item{initial_options}{PARAM_DESCRIPTION, Default: NULL}

\item{confirm}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
