% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{radiobuttons_element}
\alias{radiobuttons_element}
\title{Radio button group element}
\usage{
radiobuttons_element(id, options, initial_option = NULL, confirm = NULL)
}
\arguments{
\item{id}{character, An identifier for the action triggered
when the radio button group is changed.}

\item{options}{list, a list of block_options.}

\item{initial_option}{An option_block that exactly
matches one of the options within options, Default: NULL}

\item{confirm}{block_confirm, defines an optional
confirmation dialog after the button is clicked. Default: NULL}
}
\value{
block element class
}
\description{
A radio button group that allows a
user to choose one item from a list of possible options.
}
\details{
\itemize{
\item Works with block types: \emph{Section}, \emph{Actions}, \emph{Input}
}
}
\seealso{
Other elements: 
\code{\link{button_element}()},
\code{\link{checkbox_element}()},
\code{\link{datepicker_element}()},
\code{\link{image_element}()},
\code{\link{overflow_menu}()},
\code{\link{plaintext_element}()},
\code{\link{select_menu}()},
\code{\link{select_type_menu}()}
}
\concept{elements}
