% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{section_block}
\alias{section_block}
\title{Section Block}
\usage{
section_block(text, id = NULL, fields = NULL, accessory = NULL)
}
\arguments{
\item{text}{character, The text for the block, in the form of a text object.
Maximum length for the text in this field is 3000 characters.}

\item{id}{character, A string acting as a unique identifier for a block.
Default: NULL}

\item{fields}{An array of text objects. Any text objects included with
fields will be rendered in a compact format that allows for 2 columns
of side-by-side text. Maximum number of items is 10., Default: NULL}

\item{accessory}{One of the available element objects. Default: NULL}
}
\value{
section_block class
}
\description{
A section is one of the most flexible blocks available, it can
be used as a simple text block, in combination with text fields,
or side-by-side with any of the available block elements.
}
\details{
Available in surfaces: \emph{Modals}, \emph{Messages}, \emph{Home tabs}
}
\examples{
section_block(
  text = block_text(
    text = "A message *with some bold text* and _some italicized text_."
  )
)

section_block(
  text = block_text(
    text = "A message *with some bold text* and _some italicized text_."
  ),
  fields = wrap_blocks(
  block_text(text = 'High'),
  block_text(
    type = 'plain_text',
    text = 'string',
    emoji = TRUE)
  )
)

section_block(
  text = block_text(
    text = "*Sally* has requested you set the deadline for the Nano launch project"
  ),
  accessory = datepicker_element(
    id = 'datepicker123',
    initial_date = "1990-04-28", 
    placeholder = 'Select a date'
   )
  )

}
\seealso{
Other layout: 
\code{\link{actions_block}()},
\code{\link{as.blocks}()},
\code{\link{context_block}()},
\code{\link{divider_block}()},
\code{\link{file_block}()},
\code{\link{image_block}()},
\code{\link{input_block}()},
\code{\link{wrap_blocks}()}
}
\concept{layout}
