% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{select_menu}
\alias{select_menu}
\title{Select menu element}
\usage{
select_menu(
  id,
  placeholder,
  options,
  option_groups = NULL,
  initial_option = NULL,
  confirm = NULL,
  max_select_items = NULL
)
}
\arguments{
\item{id}{character, an identifier for the action triggered when a menu option is selected.}

\item{placeholder}{character, defines the placeholder text shown on the menu.}

\item{options}{list, a list of option objects.}

\item{option_groups}{a list of option group objects. Default: NULL}

\item{initial_option}{a single option that exactly matches one of the
options within options or option_groups, Default: NULL}

\item{confirm}{A confirm object that defines an optional
confirmation dialog that appears after a menu item is selected. Default: NULL}

\item{max_select_items}{integer, Specifies the maximum number of items that can be selected in the menu, Default: NULL}
}
\value{
block menu class
}
\description{
A select menu, just as with a standard HTML select tag,
creates a drop down menu with a list of options for a user to choose.
The select menu also includes type-ahead functionality,
where a user can type a part or all of an option string to filter the list.
}
\details{
Works with block types: \emph{Section}, \emph{Actions}, \emph{Input}

If \code{options} is specified, then \code{option_groups} should not be and vice versa.
}
\examples{
section_block(
id = 'section678',
text = 'Pick an item from the dropdown list',
accessory = select_menu(
 id = 'text1234',
 placeholder = 'Select an item',
 options = as.blocks(
   lapply(sprintf('value-\%d',1:4),block_option,
     text = '*this is plain_text text*')
   )
 )
)
}
\seealso{
Other elements: 
\code{\link{button_element}()},
\code{\link{checkbox_element}()},
\code{\link{datepicker_element}()},
\code{\link{image_element}()},
\code{\link{overflow_menu}()},
\code{\link{plaintext_element}()},
\code{\link{radiobuttons_element}()},
\code{\link{select_type_menu}()}
}
\concept{elements}
