% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{select_type_menu}
\alias{select_type_menu}
\title{Type specific select menu lists}
\usage{
select_type_menu(
  id,
  type = c("users", "conversations", "channels"),
  placeholder,
  initial = NULL,
  confirm = NULL,
  max_select_items = NULL
)
}
\arguments{
\item{id}{character, an identifier for the action triggered
when a menu option is selected.}

\item{type}{type of elements to populate,
Default: c("users", "conversations", "channels")}

\item{placeholder}{character, placeholder text shown on the menu.}

\item{initial}{initial value of the menu, must by an slack ID, Default: NULL}

\item{confirm}{A confirm object that defines an optional
confirmation dialog that appears after a menu item is selected. Default: NULL}

\item{max_select_items}{integer, Specifies the maximum number of items that can be selected in the menu, Default: NULL}
}
\value{
block menu class
}
\description{
menu lists that populate with team specific elements, such as
users, conversations and channels
}
\details{
Works with block types: \emph{Section}, \emph{Actions}, \emph{Input}
}
\seealso{
Other elements: 
\code{\link{button_element}()},
\code{\link{checkbox_element}()},
\code{\link{datepicker_element}()},
\code{\link{image_element}()},
\code{\link{overflow_menu}()},
\code{\link{plaintext_element}()},
\code{\link{radiobuttons_element}()},
\code{\link{select_menu}()}
}
\concept{elements}
