% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{wrap_blocks}
\alias{wrap_blocks}
\title{FUNCTION_TITLE}
\usage{
wrap_blocks(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
