% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{as.blocks}
\alias{as.blocks}
\title{Coerce into Blocks}
\usage{
as.blocks(x)
}
\arguments{
\item{x}{object}
}
\value{
block object
}
\description{
Turns existing object into block elements
}
\seealso{
Other layout: 
\code{\link{block_context}()},
\code{\link{block_divider}()},
\code{\link{block_section}()},
\code{\link{wrap_blocks}()}
}
\concept{layout}
