% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.blocks}
\alias{as.blocks}
\title{Coerce into Blocks}
\usage{
as.blocks(x)
}
\arguments{
\item{x}{object}
}
\value{
block object
}
\description{
Turns existing object into block elements
}
\seealso{
Other layout: 
\code{\link{actions_block}()},
\code{\link{context_block}()},
\code{\link{divider_block}()},
\code{\link{file_block}()},
\code{\link{image_block}()},
\code{\link{input_block}()},
\code{\link{section_block}()},
\code{\link{wrap_blocks}()}
}
\concept{layout}
