% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{block_button}
\alias{block_button}
\title{FUNCTION_TITLE}
\usage{
block_button(id, text, value = NULL, url = NULL, style = NULL,
  confirm = NULL)
}
\arguments{
\item{id}{PARAM_DESCRIPTION}

\item{text}{PARAM_DESCRIPTION}

\item{value}{PARAM_DESCRIPTION, Default: NULL}

\item{url}{PARAM_DESCRIPTION, Default: NULL}

\item{style}{PARAM_DESCRIPTION, Default: NULL}

\item{confirm}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
# A regular interactive button:
block_button(text = 'Click Me', value = 'click_me_123', id = 'button')

# A button with a primary style attribute:
block_button(text = 'Click Me', value = 'click_me_123', id = 'button',style = 'primary')

# A link button:

block_button(text = 'Link Button', url = 'https://api.slack.com/block-kit', id = 'button')

}
