% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositions.R
\name{block_confirm}
\alias{block_confirm}
\title{Confirmation dialog object}
\usage{
block_confirm(title, text, confirm, deny)
}
\arguments{
\item{title}{character, dialog's title}

\item{text}{block_text, the explanatory text that
appears in the confirm dialog}

\item{confirm}{character, text of the button that
confirms the action.}

\item{deny}{character, text of the button that
cancels the action.}
}
\value{
block_confirm class
}
\description{
An object that defines a dialog that provides
a confirmation step to any interactive element.
This dialog will ask the user to confirm their action
by offering a confirm and deny buttons.
}
\examples{
block_confirm(
  title = "Are you sure?",
  text = block_text("Wouldn't you prefer a good game of _chess_?"),
  confirm = "Do it",
  deny = "Stop, I've changed my mind!"
)
}
\seealso{
Other compositions: 
\code{\link{block_option_groups}()},
\code{\link{block_option}()},
\code{\link{block_text}()}
}
\concept{compositions}
