% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{block_context}
\alias{block_context}
\title{FUNCTION_TITLE}
\usage{
block_context(elements, id = NULL)
}
\arguments{
\item{elements}{An array of image elements and text objects. Maximum number of items is 10.}

\item{id}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
