% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{block_datepicker}
\alias{block_datepicker}
\title{Date picker element}
\usage{
block_datepicker(id, placeholder = NULL, initial_date = NULL, confirm = NULL)
}
\arguments{
\item{id}{An identifier for the action triggered
when a menu option is selected.}

\item{placeholder}{character, placeholder text shown on the datepicker. Default: NULL}

\item{initial_date}{Date, The initial date that is
selected when the element is loaded.
This should be in the \link[base:strftime]{strftime}
format '\%Y-\%m-\%d'. Default: NULL}

\item{confirm}{block_confirm, defines an optional
confirmation dialog after the button is clicked. Default: NULL}
}
\value{
block_datepicker class
}
\description{
An element which lets users easily
select a date from a calendar style UI.
}
\details{
\itemize{
\item Works with block types: \emph{Section}, \emph{Actions}, \emph{Input}
}
}
\examples{
block_datepicker(id = 'datepicker123',initial_date = "1990-04-28", 
placeholder = 'Select a date')
}
\seealso{
\link[parsedate:parse_date]{parse_date}
}
