% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{block_divider}
\alias{block_divider}
\title{FUNCTION_TITLE}
\usage{
block_divider(id = NULL)
}
\arguments{
\item{id}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
