% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{block_divider}
\alias{block_divider}
\title{Divider Block}
\usage{
block_divider(id = NULL)
}
\arguments{
\item{id}{character, unique identifier for a block. Default: NULL}
}
\value{
block_divider class
}
\description{
A content divider, like an hr HTML DOM,
to split up different blocks inside of a message.
}
\details{
Available in surfaces: \emph{Modals}, \emph{Messages}, \emph{Home tabs}
}
\examples{
block_divider()
}
\seealso{
Other layout: 
\code{\link{as.blocks}()},
\code{\link{block_context}()},
\code{\link{block_section}()},
\code{\link{wrap_blocks}()}
}
\concept{layout}
