% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{block_image}
\alias{block_image}
\title{FUNCTION_TITLE}
\usage{
block_image(url = NULL, alt_text = "image")
}
\arguments{
\item{url}{PARAM_DESCRIPTION, Default: NULL}

\item{alt_text}{PARAM_DESCRIPTION, Default: 'image'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
