% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositions.R
\name{block_option}
\alias{block_option}
\title{Create a block for options}
\usage{
block_option(text, value, description = NULL, url = NULL)
}
\arguments{
\item{text}{character, text that defines the text
shown in the option on the menu.}

\item{value}{character, The string value that will be
passed to your app when this option is chosen.}

\item{description}{character, text that defines a line
of descriptive text shown below the  text field beside
the radio button. Default: NULL}

\item{url}{character, A URL to load in the user's browser
when the option is clicked. Default: NULL}
}
\value{
block_option class
}
\description{
An object that represents a single selectable item
in a select menu, multi-select menu, radio button group,
or overflow menu.
}
\examples{
block_option(text = 'Maru',value = 'maru')
}
\seealso{
Other compositions: 
\code{\link{block_confirm}()},
\code{\link{block_option_groups}()},
\code{\link{block_text}()}
}
\concept{compositions}
