% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{block_section}
\alias{block_section}
\title{FUNCTION_TITLE}
\usage{
block_section(text, id = NULL, fields = NULL, accessory = NULL)
}
\arguments{
\item{text}{PARAM_DESCRIPTION}

\item{id}{PARAM_DESCRIPTION, Default: NULL}

\item{fields}{PARAM_DESCRIPTION, Default: NULL}

\item{accessory}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\concept{layout}
