% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositions.R
\name{block_text}
\alias{block_text}
\title{Create a block for text}
\usage{
block_text(text, type = c("mrkdwn", "plain-text"), emoji = NULL,
  verbatim = NULL)
}
\arguments{
\item{text}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION, Default: c("mrkdwn", "plain-text")}

\item{emoji}{PARAM_DESCRIPTION, Default: NULL}

\item{verbatim}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
