% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post_block}
\alias{post_block}
\title{Post Block Element}
\usage{
post_block(
  block,
  channel,
  thread_ts = NULL,
  token = Sys.getenv("SLACK_API_TOKEN")
)
}
\arguments{
\item{block}{block element(s)}

\item{channel}{character, Channel ID, label or a URL link to a message}

\item{thread_ts}{character slack api timestamp of the format
xxxxxxxxxx.xxxxxx, Default: NULL}

\item{token}{Slack API token, Default: Sys.getenv("SLACK_API_TOKEN")}
}
\value{
\link[httr:response]{response}
}
\description{
Post block to slack channel
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[slackcalls]{post_slack}}

Other post: 
\code{\link{post_thread}()},
\code{\link{prep_channel}()}
}
\concept{post}
