% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post_block}
\alias{post_block}
\title{FUNCTION_TITLE}
\usage{
post_block(block, channel, token = Sys.getenv("SLACK_API_TOKEN"))
}
\arguments{
\item{block}{PARAM_DESCRIPTION}

\item{channel}{PARAM_DESCRIPTION}

\item{token}{PARAM_DESCRIPTION, Default: Sys.getenv("SLACK_API_TOKEN")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[slackcalls]{post_slack}}
}
