% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post_thread}
\alias{post_thread}
\title{Post Block Element to Thread}
\usage{
post_thread(res, block)
}
\arguments{
\item{res}{\link[httr:response]{response} from \link[slackblocks:post_block]{post_block}}

\item{block}{block element(s)}
}
\value{
\link[httr:response]{response}
}
\description{
Post block to slack channel thread
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
Other post: 
\code{\link{post_block}()},
\code{\link{prep_channel}()}
}
\concept{post}
