% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{prep_channel}
\alias{prep_channel}
\title{Prepare Channel}
\usage{
prep_channel(channel, ts = NULL)
}
\arguments{
\item{channel}{character}

\item{ts}{character slack api timestamp of the format
xxxxxxxxxx.xxxxxx, Default: NULL}
}
\value{
channel
}
\description{
Function that will parse multiple types of channel inputs.
}
\details{
channel can be either
- channel label
- channel id
- link to a message from slack
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[slackteams]{validate_channel}}

Other post: 
\code{\link{post_block}()},
\code{\link{post_thread}()}
}
\concept{post}
