% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{prep_channel}
\alias{prep_channel}
\title{FUNCTION_TITLE}
\usage{
prep_channel(channel, ts = NULL)
}
\arguments{
\item{channel}{PARAM_DESCRIPTION}

\item{ts}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[slackteams]{validate_channel}}
}
