% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_blocks}
\alias{wrap_blocks}
\title{Array of blocks}
\usage{
wrap_blocks(...)
}
\arguments{
\item{...}{block elements}
}
\value{
list of blocks elements
}
\description{
wrap block class objects in a list
}
\examples{
wrap_blocks(
    image_element(
      url = 'https://image.freepik.com/free-photo/red-drawing-pin_1156-445.jpg',
      alt_text = 'images'
    ),
    block_text(
      text = 'Location: **Dogpatch**'
    )
  )
}
\seealso{
Other layout: 
\code{\link{actions_block}()},
\code{\link{as.blocks}()},
\code{\link{context_block}()},
\code{\link{divider_block}()},
\code{\link{file_block}()},
\code{\link{image_block}()},
\code{\link{input_block}()},
\code{\link{section_block}()}
}
\concept{layout}
