% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{wrap_blocks}
\alias{wrap_blocks}
\title{Array of blocks}
\usage{
wrap_blocks(...)
}
\arguments{
\item{...}{block elements}
}
\value{
list of blocks elements
}
\description{
wrap block class objects in a list
}
\examples{
wrap_blocks(
    block_image(
      url = 'https://image.freepik.com/free-photo/red-drawing-pin_1156-445.jpg',
      alt_text = 'images'
    ),
    block_text(
      text = 'Location: **Dogpatch**'
    )
  )
}
\seealso{
Other layout: 
\code{\link{as.blocks}()},
\code{\link{block_context}()},
\code{\link{block_divider}()},
\code{\link{block_section}()}
}
\concept{layout}
